requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.2
django-coverage-plugin==1.3.1
Sphinx==1.4.5
django-extensions==1.7.4
Werkzeug==0.11.11
django-test-plus==1.0.15
factory_boy==2.7.0
django-debug-toolbar==1.5

# improved REPL
ipdb==0.10.1

pytest-django==2.9.1
pytest-sugar==0.7.1

# Own 08/09/2016
pudb==2015.4.1
ipython==4.0.3
django-sslserver==0.19
